pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vishwamai"
version = "0.1.0"
description = "A Mixture-of-Experts Language Model with Multi-Level Attention"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Vishwamai Research Team", email = "team@vishwamai.org"},
]
maintainers = [
    {name = "Vishwamai Research Team", email = "team@vishwamai.org"},
]
keywords = [
    "deep-learning",
    "machine-learning",
    "natural-language-processing",
    "mixture-of-experts",
    "multi-level-attention",
    "transformer",
    "language-model",
    "tpu",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentencepiece>=0.1.99",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "einops>=0.6.1",
    "flash-attn>=2.0.0",
    "fairscale>=0.4.13",
]

[project.optional-dependencies]
tpu = [
    "torch-xla>=2.0.0",
    "cloud-tpu-client>=0.10",
]
training = [
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "accelerate>=0.21.0",
]
serving = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "pylint>=2.17.4",
]
docs = [
    "sphinx>=6.2.1",
    "sphinx-rtd-theme>=1.2.2",
    "sphinx-autodoc-typehints>=1.23.0",
]

[project.urls]
Homepage = "https://github.com/organization/vishwamai"
Documentation = "https://vishwamai.readthedocs.io/"
Repository = "https://github.com/organization/vishwamai.git"
"Bug Tracker" = "https://github.com/organization/vishwamai/issues"

[project.scripts]
vishwamai-train = "vishwamai.scripts.train_model:main"
vishwamai-eval = "vishwamai.scripts.evaluate_model:main"
vishwamai-serve = "vishwamai.scripts.serve_model:main"
vishwamai-preprocess = "vishwamai.scripts.preprocess_data:main"
vishwamai-tokenizer = "vishwamai.scripts.train_tokenizer:main"
vishwamai-export = "vishwamai.scripts.export_model:main"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=vishwamai --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["vishwamai"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
ignore_errors = true
omit = [
    "tests/*",
    "setup.py",
]

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # wrong-hanging-indentation
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "W0511",  # fixme
]

[tool.pylint.format]
max-line-length = 88

[tool.bandit]
exclude_dirs = ["tests"]
targets = ["vishwamai"]
